+++ src/core/useDepositERC20/index.ts
@@ -1,1 +1,2 @@
 export * from './useDepositERC20';
+export * from './types';
