+++ src/core/useDepositERC20/useDepositERC20.tsx
@@ -1,89 +1,56 @@
-import { useReducer } from 'react';
+import { useState } from 'react';
 import { Client } from '@aragon/sdk-client';
 import { DaoDepositSteps, DepositParams } from '@aragon/sdk-client';
 import { useMutation, UseMutationOptions } from 'react-query';
 
 import { useAragon } from '../../context';
-
-export enum DepositERC20Status {
-  IDLE = 'idle',
-  DEPOSITING = 'depositing',
-  CHECKING_ALLOWANCE = 'checkingAllowance',
-  CONFIRMING_ALLOWANCE_UPDATE = 'confirmingAllowanceUpdate',
-  CONFIRMING_DEPOSIT = 'confirmingDeposit',
-  SUCCESS = 'success',
-  ERROR = 'error',
-}
-
-export interface DepositERC20State {
-  allowance: bigint | null;
-  updateAllowanceTxid: string | null;
-  depositTxid: string | null;
-  depositStatus: DepositERC20Status;
-}
-
-const initialState: DepositERC20State = {
-  allowance: null,
-  updateAllowanceTxid: null,
-  depositTxid: null,
-  depositStatus: DepositERC20Status.IDLE,
-};
-
-function reducer(state: DepositERC20State, update: Partial<DepositERC20State>) {
-  return { ...state, ...update };
-}
+import { DepositERC20Status } from './types';
 
 export function useDepositERC20(depositParams: DepositParams, options?: UseMutationOptions) {
-  const [state, dispatch] = useReducer(reducer, initialState);
+  const [allowance, setAllowance] = useState<bigint | null>(null);
+  const [updateAllowanceTxid, setUpdateAllowanceTxid] = useState<string | null>(null);
+  const [depositTxid, setDepositTxid] = useState<string | null>(null);
+  const [depositStatus, setDepositStatus] = useState<DepositERC20Status>(DepositERC20Status.IDLE);
+
   const { baseClient: client } = useAragon();
 
   const deposit = async (client: Client, depositParams: DepositParams) => {
-    const steps = client.methods.deposit(depositParams);
-    dispatch({ depositStatus: DepositERC20Status.DEPOSITING });
-    for await (const step of steps) {
-      try {
+    try {
+      const steps = client.methods.deposit(depositParams);
+      setDepositStatus(DepositERC20Status.WAITING_FOR_SIGNER);
+      for await (const step of steps) {
         switch (step.key) {
           case DaoDepositSteps.CHECKED_ALLOWANCE:
-            dispatch({
-              allowance: step.allowance,
-              depositStatus: DepositERC20Status.CHECKING_ALLOWANCE,
-            });
-            // console.log({ state, step });
+            setDepositStatus(DepositERC20Status.CHECKING_ALLOWANCE);
+            setAllowance(step.allowance);
             break;
           case DaoDepositSteps.UPDATING_ALLOWANCE:
-            dispatch({
-              depositTxid: step.txHash,
-              depositStatus: DepositERC20Status.CONFIRMING_ALLOWANCE_UPDATE,
-            });
-            // console.log({ state, step });
+            setUpdateAllowanceTxid(step.txHash);
+            setDepositStatus(DepositERC20Status.CONFIRMING_ALLOWANCE_UPDATE);
             break;
           case DaoDepositSteps.UPDATED_ALLOWANCE:
-            dispatch({
-              allowance: step.allowance,
-            });
-            // console.log({ state, step });
+            setAllowance(step.allowance);
             break;
           case DaoDepositSteps.DEPOSITING:
-            dispatch({
-              depositTxid: step.txHash,
-              depositStatus: DepositERC20Status.CONFIRMING_DEPOSIT,
-            });
-            // console.log({ state, step });
+            setDepositTxid(step.txHash);
+            setDepositStatus(DepositERC20Status.CONFIRMING_DEPOSIT);
             break;
           case DaoDepositSteps.DONE:
-            dispatch({ depositStatus: DepositERC20Status.SUCCESS });
-            // console.log({ state, step });
+            setDepositStatus(DepositERC20Status.SUCCESS);
             break;
         }
-      } catch (err) {
-        dispatch({ depositStatus: DepositERC20Status.ERROR });
-        console.error(err);
       }
+    } catch (err) {
+      setDepositStatus(DepositERC20Status.ERROR);
+      console.error(err);
     }
   };
 
   return {
-    ...state,
+    allowance,
+    updateAllowanceTxid,
+    depositTxid,
+    depositStatus,
     ...useMutation({
       mutationKey: ['depositEth', depositParams.daoAddressOrEns],
       mutationFn: async () => await deposit(client, depositParams),
